\name{saveAffinityPlot.bgx}
\alias{saveAffinityPlot.bgx}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Save a plot of affinity categorisation.}
\description{
  This internal function saves a plot showing how probes were categorised into affinity categories.
}
\usage{
saveAffinityPlot.bgx(originalAffinities, categories, dir, probecat_threshold)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{originalAffinities}{The affinities of the probes.}
  \item{categories}{The categories of the probes.}
  \item{dir}{Name of a directory in which to save the plot.}
  \item{probecat_threshold}{The minimum number of probes per category that was used to categorise the probes.}
}
\details{
}
\value{
}
\references{See \code{\link{bgx}}}
\author{Ernest Turro}

\seealso{\code{\link{bgx}}}
\examples{
}
\keyword{internal}% at least one, from doc/KEYWORDS
